add task erb and import 2 erb files

+++ Rakefile
@@ -7,10 +7,20 @@ end
 desc 'initialize task'
 task :initialize do
   FileList['_*'].each do |src|
+    next if /\.erb$/ =~ src
     dest = "#{ENV['HOME']}/.#{src[1..-1]}"
     src, dest = File.expand_path(src), File.expand_path(dest)
     # src: _vimrc, dest: ~/.vimrc
     rm dest
     ln_s src, dest
+  end
+end
+
+desc 'erb'
+task :erb do
+  FileList['_*.erb'].each do |src|
+    dest = File.expand_path('~') + "/.#{src.match(/^_(.*?)\.erb$/)[1]}"
+    rm dest
+    sh "erb -r ./secret.rb #{src} > #{dest}"
   end
 end
